+++ pkg/nivel-F39B6ACCF1574A6E/upgrade.sql
@@ -1,7 +1,7 @@
 -- data taken from empleo.denominacion.nivel
 CREATE OR REPLACE TABLE nivel (
-    id SMALLINT UNSIGNED AUTO_INCREMENT,
-    code SMALLINT UNSIGNED,
+    id INT UNSIGNED AUTO_INCREMENT,
+    code INT UNSIGNED,
     nombre VARCHAR(100) UNIQUE NOT NULL,
 
     PRIMARY KEY pk_nivel_id(id),
